fix api update user

See merge request aibles-spring-learning/graduation-thesis/okrs/okrs-backend!63

+++ okrs-core/src/main/java/org/ptit/okrs/core/service/UserService.java
@@ -30,7 +30,7 @@ public interface UserService extends BaseService<User> {
       String userId,
       String name,
       String phone,
-      Long dateOfBirth,
+      Integer dateOfBirth,
       Gender gender,
       String address);
